Changed of argument in binary_search_recursive()
Code presented by you wasn't wrong but at the point in your video where that error occurred it was __ "Index Error: list index out of range".
It occurred because value of argument( i.e. right_index) when we called the function was given as length of list.
So whenever a number to be search was not present in the list our Algorithm would just search and search and eventually tried to go beyond available indices.
Hence just by changing value of that argument ( i.e. right_index) at calling stage of function (i.e. binary_search_recursive() )we can save the extra checkpoint that is been omitted in this above code.